iscsi: Add /etc/iscsi/iscsi.initramfs
The presence of the file /etc/iscsi/iscsi.initramfs will trigger the
open-iscsi initramfs construction hooks to copy this file inside the
initramfs as /etc/iscsi.initramfs.

At boot time (within the initramfs), the open-iscsi hooks will
evaluate the file to determine whether and how to perform an iSCSI
boot.  We use this to check for an iBFT; if found then we set
ISCSI_AUTO=1 which will then trigger the open-iscsi hooks to perform
an iBFT boot.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>