debian: Work around broken getpwuid() call in open-iscsi.finalrd
The open-iscsi.finalrd script runs an iscsid instance within the
finalrd immediately before power-off, in order to log out from the
iSCSI target.

As currently implemented, this does not work since iscsid fails on a
call to getpwuid().  This results in the system hanging for thirty
seconds on shutdown, and failing to log out from the iSCSI target.

Work around this bug by ensuring that libnss_files.so is present
within the finalrd, and creating a stub /etc/passwd if needed
containing an entry for the root user.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>